requirements.txt
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pypandoc==1.13
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
python-docx